requirements.txt
streamlit==1.30.0
langchain==0.1.9
langgraph==0.0.27
chromadb==0.4.22
python-dotenv==1.0.1
google-generativeai==0.3.2
pydantic==2.6.1
pandas==2.2.0
tqdm==4.66.2
python-docx==1.1.0
PyPDF2==3.0.1